requirements.txt
# Python 3.13.2

# retreiving and wrangling data
requests~=2.32.3
beautifulsoup4~=4.12.3
pandas~=2.2.3
numpy~=2.2.4
scikit-learn~=1.6.1

# saving the data
s3fs~=2024.12.0

# natural language processing
nltk~=3.9.1

# open AI API
openai~=1.60.1

# store keys in env file
python-dotenv~=0.21.0

# generate audio
elevenlabs~=1.55.0

# time
arrow~=1.3.0